Created an abstract Screen class to hold information for all future objects of type Screen to simplify making loading pages. LoadScreen has been simplified and extended from this abstract class.
The capability to use GIFs as loading screen images has been incorporated.
All images are located in the img resource directory. A method in Screen,
setBackgroundImage(), takes care of all details concerning resizing the
image and displaying it on screen. This means that all we need is to create
a GIF of the loading screen, place it into the resource directory and use
it and we have easy to create, attractive loading screens.

The main class was also simplified and refactored all settings of the display
into the Screen abstract class.
Signed-off-by: neferpitou <mdavila97@gmail.com>