Last night I discovered the CardLayout API, which allows us to switch between JPanels easily. With the backbone in place, all we have to do is subclass JPanel classes and worry about the layout and ActionListeners. Much easier than what I was trying to do for the last two days!
To give you an idea of how the flow of the project looks, it all starts at
the main method, which dynamically calls an object of the Screen class.
We don't care about what it is or where it is in memory since once it goes
away the game is over. That Screen instance is a subclass of JFrame and
houses a JPanel with a CardLayout. That JPanel has other JPanels added to
it (as many as we want) and then is added to the screen object (which acts
as a JFrame). There are helper methods in the Screen class to choose the
appropriate panel to display based on the tag (which should be included in
all our JPanel subclasses). Right now, it just adds one panel and that
panel is colored blue, so we can see that adding and selecting panels do
work.


Signed-off-by: neferpitou <mdavila97@gmail.com>