Game Kernel Implemented
A large amount of work has gone into reorganizing the project so that
there are as few references as possible needed to get the game running.
The biggest addition to this commit is that VGKernel was created to
handle the bulk of the video game's goals. The Screen class was
refactored into an inner class within the kernel, because only the
kernel needs to know how to work with the screen.

On a successful run, the project should display the loading screen and
the main menu, which should be blue.