Panels Refactored Into Kernel
The classes underneath the panel package have been refactored into the
kernel due to access issues. The panels need to perform functions that
involve opening and closing screens, which only the kernel has access
to. Moving these independent classes into the kernel as inner classes
simplifies access to the screen for the panels.

The ImagePanel class has largely been abstracted into another class,
LoadImage, and an etc package has been created to hold classes that do
various functions that may not fit into an already existing package. A
small change in the form of adding a method to change the background
image of the LoadingPanel was included since CardLayout does not make
removing panels easy. Therefore, there is only one loading screen and
the representation for that screen may/may not change.