Updated Documentation
The documentation in the README has been updated with the names of
everyone involved in the project and a basic overview of it, which is so
far one sentence long. I created a new documentation folder where I
included two scans of sketches I made to help the preplanning stage.

The order that panels are initialized has changed. Now, most of the
loading takes place in the thread in the loadMaterials() class and the
try-catch block that halted the thread for five seconds has been
removed. The result is that the loading screen is essentially skipped
for now.