All Panels Created
A new panel has been created which is where we will render all of the
fighting. This panel class will have the game loop. The selection screen
has been updated with an invisible count down timer that gives the user
sixty seconds to make a choice before the game starts. For now, it
prints it's contents to standard output to verify that it works.

The actual game panel has the same placeholder background as every other
panel.