HashMap Update For Kernel and Views
Views now placed into a HashMap that identifies them by a string. A new
method getView() has been created to index the hashmap and return the
panel identified by the string parameter. The kernel now implements an
interface ViewLabels which holds only variables to be able to identify
the views to be switched in all of the actionlisteners for each view.
Now views do not have to be created and destroyed every time a button is
pressed.

Kernel threading has also been updated to reflect the changes in how
views are accessed.

Parameters implementation changed from arrays to collections.