More Kernel Optimizations
The LoadImage class, used for loading images from the disk into memory
in an Image object, has been merged with the kernel.  In addition, the
ImagePanel class has been made more generic, now being a JPanel that
paints an image from disk on itself.

This allows views to load images for their backgrounds through the
kernel by creating a new ImagePanel and then retrieving the image from
that object (or using that ImagePanel).