Refined Singleton object patterns
Moved the instance of the kernel from the views into the interface. That
way, every view will automatically have access to it from the interface.

The return settings call has been removed since we don't have any yet
and when we do, we might choose a different implementation for it.