Java 8 Compliant Game Push
Code has been updated to take advantage of many facilities in Java 8.
The abstract VanillaCharacter class has been deprecated.
In the FatalView interface, the Strings identifying views have been
replaced with an enumerator.
In the Character interface, the array determining the types of positions
a character can be in has been replaced with an enumerator.
Anonymous classes in the kernel have been replaced with lambda
expressions.