Stage Background is animated
The stage background is now animated.

This was achieved by removing the calls to load an image in the kernel
with a call to load an image into an ImageIcon and then extracting the
image. While images returned from either Toolkit or ImageIO is static,
one obtained from an ImageIcon is animated because a MediaTracker is
automatically started on it.