Sketched out code for in-game characters
I determined that having a character interface and a character abstract
class was redundant, and that it would be easier to work with a class
than an interface, so the interface has been scrapped. All code has been
moved into VanillaCharacter and fleshed out, but the code to perform
specific actions is not present.

Also created a factory class to make constructing characters and stages
easy.